+++ pkg/costattribution/active_tracker.go
@@ -18,6 +18,8 @@ import (
 	"github.com/grafana/mimir/pkg/costattribution/costattributionmodel"
 )
 
+const trackedSeriesFactor = 2
+
 type counters struct {
 	activeSeries           atomic.Int64
 	nativeHistograms       atomic.Int64
@@ -27,6 +29,7 @@ type counters struct {
 type ActiveSeriesTracker struct {
 	userID                                         string
 	activeSeriesPerUserAttribution                 *prometheus.Desc
+	attributedOverflowLabels                       *prometheus.Desc
 	activeNativeHistogramSeriesPerUserAttribution  *prometheus.Desc
 	activeNativeHistogramBucketsPerUserAttribution *prometheus.Desc
 	logger                                         log.Logger
@@ -34,6 +37,8 @@ type ActiveSeriesTracker struct {
 	labels         []costattributionmodel.Label
 	overflowLabels []string
 
+	// maxCardinality is the cardinality at which tracker enters the overflow state.
+	// There are up to trackedSeriesFactor*maxCardinality series in observed map.
 	maxCardinality   int
 	cooldownDuration time.Duration
 
@@ -73,6 +78,9 @@ func NewActiveSeriesTracker(userID string, trackedLabels []costattributionmodel.
 	ast.activeSeriesPerUserAttribution = prometheus.NewDesc("cortex_ingester_attributed_active_series",
 		"The total number of active series per user and attribution.", variableLabels[:len(variableLabels)-1],
 		prometheus.Labels{trackerLabel: defaultTrackerName})
+	ast.attributedOverflowLabels = prometheus.NewDesc("cortex_attributed_series_overflow_labels",
+		"The overflow labels for this tenant. This metric is always 1 for tenants with active series, it is only used to have the overflow labels available in the recording rules without knowing their names.", variableLabels[:len(variableLabels)-1],
+		prometheus.Labels{trackerLabel: defaultTrackerName})
 	ast.activeNativeHistogramSeriesPerUserAttribution = prometheus.NewDesc("cortex_ingester_attributed_active_native_histogram_series",
 		"The total number of active native histogram series per user and attribution.", variableLabels[:len(variableLabels)-1],
 		prometheus.Labels{trackerLabel: defaultTrackerName})
@@ -98,9 +106,12 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 	defer bufferPool.Put(buf)
 	at.fillKeyFromLabels(lbls, buf)
 
+	// Happy path, maybe we're already tracking this cost-attribution labels combination.
+	// First try just with a read lock.
 	at.observedMtx.RLock()
 	c, ok := at.observed[string(buf.Bytes())]
 	if ok {
+		// We have this combination already tracked, just increase the counter for it.
 		c.activeSeries.Inc()
 		if nativeHistogramBucketNum >= 0 {
 			c.nativeHistograms.Inc()
@@ -110,34 +121,15 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 		return
 	}
 
-	if !at.overflowSince.IsZero() {
-		at.observedMtx.RUnlock()
+	// We don't have this combination tracked.
+	// If we are at capacity, there's no need to grab the write lock to try to add a new series.
+	if len(at.observed) >= trackedSeriesFactor*at.maxCardinality {
 		at.overflowCounter.activeSeries.Inc()
 		if nativeHistogramBucketNum >= 0 {
 			at.overflowCounter.nativeHistograms.Inc()
 			at.overflowCounter.nativeHistogramBuckets.Add(int64(nativeHistogramBucketNum))
 		}
-		return
-	}
-
-	c, ok = at.observed[string(buf.Bytes())]
-	if ok {
-		c.activeSeries.Inc()
-		if nativeHistogramBucketNum >= 0 {
-			c.nativeHistograms.Inc()
-			c.nativeHistogramBuckets.Add(int64(nativeHistogramBucketNum))
-		}
 		at.observedMtx.RUnlock()
-		return
-	}
-
-	if !at.overflowSince.IsZero() {
-		at.observedMtx.RUnlock()
-		at.overflowCounter.activeSeries.Inc()
-		if nativeHistogramBucketNum >= 0 {
-			at.overflowCounter.nativeHistograms.Inc()
-			at.overflowCounter.nativeHistogramBuckets.Add(int64(nativeHistogramBucketNum))
-		}
 		return
 	}
 	at.observedMtx.RUnlock()
@@ -145,6 +137,7 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 	at.observedMtx.Lock()
 	defer at.observedMtx.Unlock()
 
+	// Are we alredy tracking this series?
 	c, ok = at.observed[string(buf.Bytes())]
 	if ok {
 		c.activeSeries.Inc()
@@ -155,7 +148,8 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 		return
 	}
 
-	if !at.overflowSince.IsZero() {
+	// If we can't store more series, we count this series in the overflow counter.
+	if len(at.observed) >= trackedSeriesFactor*at.maxCardinality {
 		at.overflowCounter.activeSeries.Inc()
 		if nativeHistogramBucketNum >= 0 {
 			at.overflowCounter.nativeHistograms.Inc()
@@ -164,16 +158,7 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 		return
 	}
 
-	if len(at.observed) >= at.maxCardinality {
-		at.overflowSince = now
-		at.overflowCounter.activeSeries.Inc()
-		if nativeHistogramBucketNum >= 0 {
-			at.overflowCounter.nativeHistograms.Inc()
-			at.overflowCounter.nativeHistogramBuckets.Add(int64(nativeHistogramBucketNum))
-		}
-		return
-	}
-
+	// We still can store more series, add it to the map.
 	counter := &counters{}
 	counter.activeSeries.Inc()
 	if nativeHistogramBucketNum >= 0 {
@@ -181,6 +166,11 @@ func (at *ActiveSeriesTracker) Increment(lbls labels.Labels, now time.Time, nati
 		counter.nativeHistogramBuckets.Add(int64(nativeHistogramBucketNum))
 	}
 	at.observed[string(buf.Bytes())] = counter
+
+	// Check whether we entered the overflow state.
+	if len(at.observed) > at.maxCardinality && at.overflowSince.IsZero() {
+		at.overflowSince = now
+	}
 }
 
 func (at *ActiveSeriesTracker) Decrement(lbls labels.Labels, nativeHistogramBucketNum int) {
@@ -213,9 +203,6 @@ func (at *ActiveSeriesTracker) Decrement(lbls labels.Labels, nativeHistogramBuck
 		at.observedMtx.Unlock()
 		return
 	}
-	at.observedMtx.RUnlock()
-
-	at.observedMtx.RLock()
 	defer at.observedMtx.RUnlock()
 
 	if !at.overflowSince.IsZero() {
@@ -230,6 +217,8 @@ func (at *ActiveSeriesTracker) Decrement(lbls labels.Labels, nativeHistogramBuck
 }
 
 func (at *ActiveSeriesTracker) Collect(out chan<- prometheus.Metric) {
+	out <- prometheus.MustNewConstMetric(at.attributedOverflowLabels, prometheus.GaugeValue, 1, at.overflowLabels[:len(at.overflowLabels)-1]...)
+
 	at.observedMtx.RLock()
 	if !at.overflowSince.IsZero() {
 		var activeSeries int64