+++ pkg/ingester/reactivelimiter.go
@@ -5,6 +5,8 @@ package ingester
 import (
 	"context"
 
+	"github.com/failsafe-go/failsafe-go/adaptivelimiter"
+	"github.com/failsafe-go/failsafe-go/priority"
 	"github.com/go-kit/log"
 	"github.com/grafana/dskit/services"
 	"github.com/prometheus/client_golang/prometheus"
@@ -17,63 +19,29 @@ const (
 	reactiveLimiterRequestTypeLabel = "request_type"
 )
 
-type rejectionPrioritizer struct {
-	cfg *reactivelimiter.RejectionPrioritizerConfig
-	reactivelimiter.Prioritizer
-}
-
 // The ingester reactive limiter consists of two limiters: one for push requests and one for read requests.
-// It also includes a rejectionPrioritizer which, based on recent latencies in both limiters, decides the priority of requests to reject.
+// It also includes a Prioritizer which, based on recent latencies in both limiters, decides the priority of requests to reject.
 type ingesterReactiveLimiter struct {
 	service services.Service
 
-	prioritizer *rejectionPrioritizer
-	push        reactivelimiter.BlockingLimiter
-	read        reactivelimiter.BlockingLimiter
+	prioritizer *reactivelimiter.Prioritizer
+	push        reactivelimiter.ReactiveLimiter
+	read        reactivelimiter.ReactiveLimiter
 }
 
 // Returns an ingesterReactiveLimiter that uses reactivelimiter.PriorityLimiters with a Prioritizer when push and read
 // limiting is enabled, else that uses reactivelimiter.BlockingLimiter if only one of these is enabled.
-func newIngesterReactiveLimiter(prioritizerConfig *reactivelimiter.RejectionPrioritizerConfig, pushConfig *reactivelimiter.Config, readConfig *reactivelimiter.Config, logger log.Logger, registerer prometheus.Registerer) *ingesterReactiveLimiter {
-	var prioritizer *rejectionPrioritizer
-	var pushLimiter reactivelimiter.BlockingLimiter
-	var readLimiter reactivelimiter.BlockingLimiter
-
+func newIngesterReactiveLimiter(prioritizerConfig *reactivelimiter.PrioritizerConfig, pushConfig *reactivelimiter.Config, readConfig *reactivelimiter.Config, logger log.Logger, registerer prometheus.Registerer) *ingesterReactiveLimiter {
+	var prioritizer *reactivelimiter.Prioritizer
 	if pushConfig.Enabled && readConfig.Enabled {
-		// Create prioritizer to prioritize the rejection threshold between push and read limiters
-		prioritizer = &rejectionPrioritizer{
-			cfg:         prioritizerConfig,
-			Prioritizer: reactivelimiter.NewPrioritizer(logger),
-		}
-
-		// Capture rejection metrics from the prioritizer
-		promauto.With(registerer).NewGaugeFunc(prometheus.GaugeOpts{
-			Name: "cortex_ingester_rejection_rate",
-			Help: "The prioritized rate at which requests should be rejected.",
-		}, func() float64 {
-			return prioritizer.RejectionRate()
-		})
-		promauto.With(registerer).NewGaugeFunc(prometheus.GaugeOpts{
-			Name: "cortex_ingester_rejection_threshold",
-			Help: "The priority threshold below which requests should be rejected.",
-		}, func() float64 {
-			return float64(prioritizer.RejectionThreshold())
-		})
-
-		// Create limiters that use prioritizer
-		pushLimiter = newPriorityLimiter(pushConfig, prioritizer, reactivelimiter.PriorityHigh, "push", logger, registerer)
-		readLimiter = newPriorityLimiter(readConfig, prioritizer, reactivelimiter.PriorityLow, "read", logger, registerer)
-	} else {
-		pushLimiter = newBlockingLimiter(pushConfig, "push", logger, registerer)
-		readLimiter = newBlockingLimiter(readConfig, "read", logger, registerer)
+		prioritizer = reactivelimiter.NewPrioritizer(logger)
+		registerPrioritizerMetrics(prioritizer, registerer)
 	}
-
 	limiter := &ingesterReactiveLimiter{
 		prioritizer: prioritizer,
-		push:        pushLimiter,
-		read:        readLimiter,
+		push:        newReactiveLimiter(pushConfig, "push", logger, registerer, priority.Low, prioritizer),
+		read:        newReactiveLimiter(readConfig, "read", logger, registerer, priority.High, prioritizer),
 	}
-
 	if prioritizer != nil {
 		limiter.service = services.NewTimerService(prioritizerConfig.CalibrationInterval, nil, limiter.update, nil)
 	}
@@ -86,45 +54,38 @@ func (l *ingesterReactiveLimiter) update(_ context.Context) error {
 	return nil
 }
 
-func newBlockingLimiter(cfg *reactivelimiter.Config, requestType string, logger log.Logger, registerer prometheus.Registerer) reactivelimiter.BlockingLimiter {
+func newReactiveLimiter(cfg *reactivelimiter.Config, requestType string, logger log.Logger, registerer prometheus.Registerer, requestPriority priority.Priority, prioritizer *reactivelimiter.Prioritizer) reactivelimiter.ReactiveLimiter {
 	if !cfg.Enabled {
 		return nil
 	}
 
-	limiter := reactivelimiter.NewBlockingLimiter(cfg, log.With(logger, "request_type", requestType))
-	registerReactiveLimiterMetrics(limiter, requestType, registerer)
+	var limiter reactivelimiter.ReactiveLimiter
+	if prioritizer != nil {
+		limiter = reactivelimiter.NewWithPriority(cfg, logger, requestPriority, prioritizer)
+	} else {
+		limiter = reactivelimiter.New(cfg, logger)
+	}
+
+	registerReactiveLimiterMetrics(limiter.Metrics(), requestType, registerer)
 	return limiter
 }
 
-// A limiter that acquires permits for a specific priority.
-type priorityLimiter struct {
-	reactivelimiter.PriorityLimiter
-	priority reactivelimiter.Priority
+func registerPrioritizerMetrics(prioritizer *reactivelimiter.Prioritizer, r prometheus.Registerer) {
+	promauto.With(r).NewGaugeFunc(prometheus.GaugeOpts{
+		Name: "cortex_ingester_rejection_rate",
+		Help: "The prioritized rate at which requests should be rejected.",
+	}, func() float64 {
+		return prioritizer.RejectionRate()
+	})
+	promauto.With(r).NewGaugeFunc(prometheus.GaugeOpts{
+		Name: "cortex_ingester_rejection_threshold",
+		Help: "The priority threshold below which requests should be rejected.",
+	}, func() float64 {
+		return float64(prioritizer.RejectionThreshold())
+	})
 }
 
-// Returns a BlockingLimiter that uses a fixed priority to threshold all requests against the limiter.
-func newPriorityLimiter(cfg *reactivelimiter.Config, prioritizer reactivelimiter.Prioritizer, priority reactivelimiter.Priority, requestType string, logger log.Logger, registerer prometheus.Registerer) reactivelimiter.BlockingLimiter {
-	if !cfg.Enabled || prioritizer == nil {
-		return nil
-	}
-
-	limiter := reactivelimiter.NewPriorityLimiter(cfg, prioritizer, log.With(logger, "requestType", requestType))
-	registerReactiveLimiterMetrics(limiter, requestType, registerer)
-	return &priorityLimiter{
-		PriorityLimiter: limiter,
-		priority:        priority,
-	}
-}
-
-func (l *priorityLimiter) CanAcquirePermit() bool {
-	return l.PriorityLimiter.CanAcquirePermit(l.priority)
-}
-
-func (l *priorityLimiter) AcquirePermit(ctx context.Context) (reactivelimiter.Permit, error) {
-	return l.PriorityLimiter.AcquirePermit(ctx, l.priority)
-}
-
-func registerReactiveLimiterMetrics(limiterMetrics reactivelimiter.Metrics, requestType string, r prometheus.Registerer) {
+func registerReactiveLimiterMetrics(limiterMetrics adaptivelimiter.Metrics, requestType string, r prometheus.Registerer) {
 	promauto.With(r).NewGaugeFunc(prometheus.GaugeOpts{
 		Name:        "cortex_ingester_reactive_limiter_inflight_limit",
 		Help:        "Ingester reactive limiter inflight request limit.",
@@ -140,10 +101,10 @@ func registerReactiveLimiterMetrics(limiterMetrics reactivelimiter.Metrics, requ
 		return float64(limiterMetrics.Inflight())
 	})
 	promauto.With(r).NewGaugeFunc(prometheus.GaugeOpts{
-		Name:        "cortex_ingester_reactive_limiter_blocked_requests",
-		Help:        "Ingester reactive limiter blocked requests.",
+		Name:        "cortex_ingester_reactive_limiter_queued_requests",
+		Help:        "Ingester reactive limiter queued requests.",
 		ConstLabels: map[string]string{reactiveLimiterRequestTypeLabel: requestType},
 	}, func() float64 {
-		return float64(limiterMetrics.Blocked())
+		return float64(limiterMetrics.Queued())
 	})
 }