+++ pkg/streamingpromql/optimize/ast/collapse_constants_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package ast
+package ast_test
 
 import (
 	"context"
@@ -8,6 +8,8 @@ import (
 
 	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/mimir/pkg/streamingpromql/optimize/ast"
 )
 
 func TestCollapseConstants(t *testing.T) {
@@ -51,15 +53,12 @@ func TestCollapseConstants(t *testing.T) {
 		`"abc"`:         `"abc"`,
 	}
 
-	collapseConstants := &CollapseConstants{}
+	ctx := context.Background()
+	collapseConstants := &ast.CollapseConstants{}
 
 	for input, expected := range testCases {
 		t.Run(input, func(t *testing.T) {
-			expr, err := parser.ParseExpr(input)
-			require.NoError(t, err)
-
-			result, err := collapseConstants.Apply(context.Background(), expr)
-			require.NoError(t, err)
+			result := runASTOptimizationPassWithoutMetrics(t, ctx, input, collapseConstants)
 			require.Equal(t, expected, result.String())
 
 			// Check for unnecessary unary expressions.